packages/gooddata-flight-server/pyproject.toml
# (C) 2025 GoodData Corporation
[project]
name = "gooddata-flight-server"
version = "1.53.0"
description = "Flight RPC server to host custom functions"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GoodData", email = "support@gooddata.com"}
]
keywords = [
    "gooddata",
    "flight",
    "rpc",
    "flight rpc",
    "custom functions",
    "analytics",
    "headless",
    "business",
    "intelligence",
    "headless-bi",
    "cloud",
    "native",
    "semantic",
    "layer",
    "sql",
    "metrics",
]
requires-python = ">=3.9.0"
dependencies = [
    "dynaconf>=3.1.11,<4.0.0",
    "opentelemetry-api>=1.24.0,<=2.0.0",
    "opentelemetry-sdk>=1.24.0,<=2.0.0",
    "orjson>=3.8.5,<4.0.0",
    "prometheus-client~=0.20.0",
    "pyarrow>=16.1.0",
    "readerwriterlock~=1.0.9",
    "structlog>=24.0.0,<25.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Typing :: Typed",
]

[project.urls]
Documentation = "https://gooddata-flight-server.readthedocs.io/en/v1.53.0"
Source = "https://github.com/gooddata/gooddata-python-sdk"

[project.scripts]
gooddata-flight-server = "gooddata_flight_server.cli:main"

[dependency-groups]
test = [
    "pytest~=8.3.4",
    "pytest-cov~=6.0.0",
]
type = [
    "mypy~=1.11.2",
    "pydantic~=2.9.2"
]

[tool.hatch.build.targets.wheel]
packages = ["src/gooddata_flight_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
